requirements.txt
Flask
flask-cors
python-dotenv
pandas
openpyxl